+++ server/mentionscrawler/blueprints/mention.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+from ..authentication.authenticate import authenticate, enforce_json
+
+mention_bp = Blueprint("mentions", __name__, url_prefix="/")
+
+@mention_bp.route("/mentions")
