+++ server/mentionscrawler/crawlers/__init__.py
@@ -1,0 +1,5 @@
+from . import reddit
+
+
+def reddit_search(user):
+    reddit.search(user)
